+++ packages/zowe-explorer/src/trees/shared/SharedInit.ts
@@ -10,7 +10,7 @@
  */
 
 import * as vscode from "vscode";
-import { FileManagement, IZoweTree, IZoweTreeNode, Validation, ZosEncoding, ZoweScheme } from "@zowe/zowe-explorer-api";
+import { FileManagement, Gui, IZoweTree, IZoweTreeNode, Validation, ZosEncoding, ZoweScheme, imperative } from "@zowe/zowe-explorer-api";
 import { SharedActions } from "./SharedActions";
 import { SharedHistoryView } from "./SharedHistoryView";
 import { SharedTreeProviders } from "./SharedTreeProviders";
@@ -330,6 +330,33 @@ export class SharedInit {
                 ZoweExplorerApiRegister.getInstance().onProfilesUpdateEmitter.fire(Validation.EventType.UPDATE);
             });
         });
+
+        try {
+            const zoweWatcher = imperative.EventOperator.getWatcher().subscribeUser(imperative.ZoweUserEvents.ON_VAULT_CHANGED, async () => {
+                ZoweLogger.info(vscode.l10n.t("Changes in the credential vault detected, refreshing Zowe Explorer."));
+                await ProfilesUtils.readConfigFromDisk();
+                await SharedActions.refreshAll();
+                ZoweExplorerApiRegister.getInstance().onVaultUpdateEmitter.fire(Validation.EventType.UPDATE);
+            });
+            context.subscriptions.push(new vscode.Disposable(zoweWatcher.close.bind(zoweWatcher)));
+        } catch (err) {
+            Gui.errorMessage("Unable to watch for vault changes. " + JSON.stringify(err));
+        }
+
+        try {
+            const zoweWatcher = imperative.EventOperator.getWatcher().subscribeShared(
+                imperative.ZoweSharedEvents.ON_CREDENTIAL_MANAGER_CHANGED,
+                async () => {
+                    ZoweLogger.info(vscode.l10n.t("Changes in credential management detected, refreshing Zowe Explorer."));
+                    await ProfilesUtils.getProfileInfo();
+                    await SharedActions.refreshAll();
+                    ZoweExplorerApiRegister.getInstance().onCredMgrUpdateEmitter.fire(Validation.EventType.UPDATE);
+                }
+            );
+            context.subscriptions.push(new vscode.Disposable(zoweWatcher.close.bind(zoweWatcher)));
+        } catch (err) {
+            Gui.errorMessage("Unable to watch for credential manager changes. " + JSON.stringify(err));
+        }
     }
 
     public static initSubscribers(context: vscode.ExtensionContext, theProvider: IZoweTree<IZoweTreeNode>): void {